Reformatted the entirety of the code because we can't organise for shit.
