--- Frontend/src/loc-work/styles.css
@@ -1,284 +1,0 @@
-* {
-  margin: 0;
-  padding: 0;
-  box-sizing: border-box;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-header {
-}
-
-.navbar {
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  background: #3f8af4;
-  padding: 2rem 0;
-}
-
-.logo-search-container {
-  display: flex;
-  flex-grow: 1;
-}
-
-.logo-search-container img {
-  width: 6rem;
-}
-
-.search {
-  width: max-content;
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  padding: 1rem 1rem 1rem 0;
-  border-radius: 5px;
-  background: #f6f6f6;
-  flex-grow: 1;
-  margin-right: 1rem;
-}
-
-.search form {
-  width: 100%;
-  margin: 0 2rem;
-}
-
-.search input {
-  font-size: 1.2rem;
-  font-family: Arial, Helvetica, sans-serif;
-  color: #333333;
-  outline: none;
-  border: none;
-  background: transparent;
-  padding-left: 1rem;
-  flex-grow: 1;
-  width: 100%;
-}
-
-.search i {
-  font-size: 1.2rem;
-}
-
-.user-options {
-  display: flex;
-  justify-content: space-between;
-  gap: 1.5rem;
-  margin-right: 1.5rem;
-  font-size: 1.2rem;
-}
-
-.user-options a {
-  text-decoration: none;
-  color: black;
-}
-
-.hero {
-  display: flex;
-  justify-content: space-around;
-  align-items: center;
-  padding: 1rem;
-  flex-wrap: wrap;
-  gap: 1rem;
-}
-
-.hero img {
-  flex-grow: 2;
-  height: 100%;
-  max-width: 100%;
-  border-radius: 1.5rem;
-}
-
-.side-navigation {
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: flex-start;
-  padding-right: 10rem;
-  background: #f6f6f6;
-  border-radius: 1.5rem;
-
-
-}
-
-.side-navigation li {
-  list-style-type: none;
-  padding: 0.5rem 2rem;
-  line-height: 1.5rem;
-  /* min-width: 200px; */
-}
-
-.side-navigation li a {
-  text-decoration: none;
-  color: black;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-.trending-discover {
-  display: flex;
-  justify-content: space-evenly;
-  margin: 1rem 0 1rem 0;
-  flex-wrap: wrap;
-}
-
-.trending,
-.discover {
-  display: flex;
-  align-items: center;
-  justify-content: center;
-  width: 33rem;
-  height: 20rem;
-  border-radius: 2rem;
-  background: #6179b7;
-  margin-bottom: 2rem;
-}
-
-.trending h1,
-.discover h1 {
-  color: black;
-  font-size: 3rem;
-  transform: rotate(45deg);
-}
-
-.product-container {
-  display: grid;
-  grid-template-columns: repeat(auto-fill, minmax(250px, 1fr));
-  row-gap: 2rem;
-  column-gap: 2rem;
-  /* background-color: #2196F3; */
-  /* padding: 0 5rem; */
-  margin: 0 2rem;
-}
-.product-card {
-  background: #f5f5f5;
-  padding: 0.5rem 0.5rem;
-  border-radius: 0.5rem;
-  margin: 0 auto;
-}
-
-.product-image {
-  max-width: 240px;
-  /* max-height: 100%; */
-  /* object-fit: cover; */
-}
-
-.image-container {
-  background: #3f8af4;
-  position: relative;
-  border-radius: 5px;
-}
-
-.image-container .wishlist{
-  position: absolute;
-  top: 10px;
-  right: 10px;
-  color: black;
-  padding: 0.5rem;
-  font-size: 2rem;
-  background: white;
-  border-radius: 50%;
-}
-
-.image-container:hover .add-to-cart {
-  display: block; /* Make the button visible on hover */
-}
-
-.add-to-cart {
-  width: 100%;
-  padding: 1rem 0;
-  display: none; 
-  font-size: 1rem; 
-  position: absolute; 
-  bottom: 0; 
-  left: 0;
-  background-color: #333333; 
-  color: white; 
-  text-align: center;
-  border-radius: 5px;
-  font-family: Arial, Helvetica, sans-serif;
-}
-
-.product-info {
-  line-height: 2rem;
-  margin-top: 1rem;
-}
-
-.view-all {
-  display: flex;
-  justify-content: center;
-}
-
-.btn {
-  padding: 1rem 2.5rem;
-  border: none;
-  width: 20%;
-  margin: 2rem 0;
-  background: #6179b7;
-  color: white;
-  border-radius: 5px;
-}
-
-/* MEDIA QUERIES */
-
-@media (max-width: 768px) {
-  .navbar {
-    flex-direction: column;
-    gap: 1rem;
-  }
-
-  .logo-search-container {
-    flex-direction: column;
-    align-items: center;
-  }
-
-  .search {
-    width: 100%;
-  }
-
-  .side-navigation {
-    margin-bottom: 1rem;
-  }
-
-  .trending-discover {
-    padding: 0 1rem;
-    margin-top: 0;
-    gap: 1rem;
-  }
-
-
-  .trending,
-  .discover {
-    width: 100%;
-    margin: 0;
-  }
-
-  .btn {
-    width: 50%;
-  }
-}
-
-@media (max-width: 576px) {
-  .logo-search-container img {
-    display: none;
-  }
-
-  .logo-search-container .search {
-    width: auto;
-
-  }
-
-  .search input {
-    padding: 0;
-  }
-
-  .search{
-    margin: 0;
-  }
-
-  .hero {
-    gap: 0;
-  }
-
-
-  .btn {
-    width: 65%;
-  }
-}
