+++ Toivoa_Frontend/src/components/loc-work/ProductDetails.css
