+++ Toivoa_Frontend/src/images/main_logo.png
